// frontend/src/api/generated/models/catalogResponseDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Lammah Game API
 * Runtime HTTP contract for Lammah game administration, content, and gameplay.
 * OpenAPI spec version: 1.0.0
 */
import type { CatalogLocalizedTextResponseDto } from "./catalogLocalizedTextResponseDto";
import type { CatalogBannerResponseDto } from "./catalogBannerResponseDto";

export interface CatalogResponseDto {
  _id: string;
  name: CatalogLocalizedTextResponseDto;
  description?: CatalogLocalizedTextResponseDto;
  slug: string;
  banner?: CatalogBannerResponseDto;
  icon?: string;
  isActive: boolean;
  sortOrder: number;
  createdAt?: string;
  updatedAt?: string;
}